requirements.txt
# Core NLP Libraries
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1

# Semantic Similarity (Optional but recommended)
sentence-transformers>=2.2.2

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# CLI Framework
click>=8.1.7

# Document Processing
python-docx>=1.1.0
fpdf2>=2.7.6

# Text Processing
textstat>=0.7.3
language-tool-python>=2.7.1

# Utilities
requests>=2.31.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
